exported TheWitcherTab in TWCreativeModeTab improved player capability syncing tweaked some tags added sign strength get/set command

+++ src/main/java/com/barion/the_witcher/client/TWPlayerEnergyData.java
@@ -1,13 +1,13 @@
 package com.barion.the_witcher.client;
 
 public class TWPlayerEnergyData {
-    public static int energy = 0;
+    public static float energy = 0;
     public static int maxEnergy = 100;
 
-    public static void set(int energy) {TWPlayerEnergyData.energy = energy;}
+    public static void set(float energy) {TWPlayerEnergyData.energy = energy;}
     public static void setMax(int maxEnergy) {TWPlayerEnergyData.maxEnergy = maxEnergy;}
 
-    public static int get() {return energy;}
+    public static float get() {return energy;}
     public static int getMax() {return maxEnergy;}
-    public static double getPercent() {return (double) get()/getMax();}
+    public static float getPercent() {return get()/getMax();}
 }
